objectstore: store tree's containing dir
Currently the object in the object store is exactly the tree, however, in
the future we want to store metadata together with the tree. For this reason
store each object as a directory containing the tree, rather than the tree itself.

This should not affect any consumers of the API, except that objects stored
before this change are no longer accessible.

During creation of a tree, it is created in a randomly named _workdir. Rather
than moving the tree into the store, move the whole _workdir there. To refelct
this new semantics, rename `store_tree()` to `store_object()`.