objectstore: store results with tree
A pipeline produces a tree and a results dict. The tree is saved in the store, but
the results are not. This means the behaviour changes depending on whether or
not the tree was cached. As a result we currently have to be careful about where
we use caching.

This is a simple bug fix that improves the situation (but see note about. When
a tree is committed to the store the results dict is saved too and when we get()
a tree from the store the results dict is returned too.

In the future we want to restructure how we handle results, but this does not
change the behaviour beyond the caching.