sources: introduction of the SourceError
During a downloading a, if an element cannot be properly downloaded, the
program has now the possibility to send a SourceError as a signal to
its in-osbuild counterpart. Between the two the message is serialized as
JSON. Once its unpacked, the Source (client) will properly raised the
SourceError kind to propagate the information to the rest of the osbuild
stack.