fmt: separate manifest validation from result fmt
This commit introduces two kinds of formats, IN and OUT ones.

IN formats are responsible of parsing manifests, validate them, and
transform them to python objects.
OUT formats are responsible of formatting the result into something and
to print it.

A new option is added: --result-format which let the user decide how to
best tailor its result formatting. The --json flag is now a proxy for
the new option. No --json flag means --result-format=text and a a --json
flag means --result-format=json/auto. Setting a manual --result-format
always override the --json one.