stages/rpm: support marking install as ostree
An ostree system can be identified as such by the presence of a
marker file: /run/ostree-booted. The `rpm-ostree` tool also
creates this marker during the _installation_ of the system[1,2].
Recently, starting with F36, the authselect package has taken
has become mandatory[3] and is now owning the nsswitch config.
An rpm-ostree system, which has parts of the user database in
/usr, release on the nss-altfiles to read these databases. The
necessary entries are added during the post-processing, which
is called in our `org.osbuild.ostree.preptree` stage[4]. This
installation is skipped though if the nsswitch is the file is
a symlink, indicating that it is owned by some other package,
like authselect. So the F36 authselect change first broke rpm-
ostree[6]. The fix was to check for `/run/ostree-booted` in
the authselect scriptlet and special case this situation[7,8].
Now, our `org.osbuild.rpm` stage does not yet have the ability
to pretend it is a running ostree system and thus we did not
get the special treatment resulting in nss-altfiles not being
enabled in our ostree commits. Therefore the passwd database in
/usr was not read and a lot of daemons and programs without a
valid user, like e.g. `sshd`.
This change introduces a new option, `ostree_booted` that if
set, will create the `/run/ostree-booted` marker and thus our
installation phase will get the same treatments from packages
as rpm-ostree. Hopefully.

[1] https://github.com/coreos/rpm-ostree/blob/730bec87b129354ce2f43569fd5257bf249e63d1/rust/src/builtins/compose/mod.rs#L24
[2] https://github.com/coreos/rpm-ostree/blob/6211d1452e1c2c6cf3187d638d81c5f4e0d7c9c3/src/app/rpmostree-compose-builtin-tree.cxx#L501
[3] https://fedoraproject.org/wiki/Changes/Make_Authselect_Mandatory
[4] https://github.com/coreos/rpm-ostree/blob/7993c6f5657062c1502f74f03e8e267066ca87bc/rust/src/composepost.rs#L635
[5] https://src.fedoraproject.org/rpms/authselect/blob/d614caeca1a68f55542aefd0d76bda2691c85d24/f/0010-spec-fix-detection-of-ostree-system.patch
[6] https://bugzilla.redhat.com/show_bug.cgi?id=2034360
[7] https://src.fedoraproject.org/rpms/authselect/c/d614caeca1a68f55542aefd0d76bda2691c85d24?branch=f36
[8] https://github.com/authselect/authselect/issues/48