library: the first oxidization of osbuild
This provides the new `/library` path in the `osbuild` repository where
libraries live that can be shared amongst `osbuild` and its modules (and
possibly others that need `osbuild`-functionality).

These libraries are written in the Rust programming language and might
(or might not) have bindings for a specific language.

In this initial implementation we make
`osbuild::utility::filesystem::copy_tree` which is implemented in
Rust available through a `osbuild_py` Python package for direct use in
`osbuild`.

This allows to not need to copy the `cp` executable into the buildroot
and giving it `CAP_MAC_ADMIN`.

Most of the work is figuring out the actual packaging of this module for
multiple architectures.