spec: split out containers sub-package
Split out all inputs, sources and stages that are needed to embed
container images into a new `osbuild-containers` subpackage. This
is mostly because all of them are dependent on `skopeo` and since
this is also true for the source, the host osbuild will also need
to depend on it. Add this dependency for the new `containers` sub
package (and only there).