util/fscache: allow directory reclaim
Since we store lock-files in the directory they lock, we already have
to gracefully retry on lock-creation if another process raced us before
we could claim the lock in the directory we created. However, we did not
allow for maintenance tasks to claim entries if they lack a lockfile.
This, however, is necessary if we consider file-system crashes and thus
face corrupted entries that possibly lack a lockfile but are non-empty.
Note that this cannot happen without any corruption to have occurred, so
this cannot interrupt normal operation, especially if the cache is
volatile.

Extend the cache-handling to allow directory reclaim. This means, during
directory creation, we allow racing claims and gracefully retry the
operation. During deletion we assumed we could always drop empty
directories we own. Extend this to gracefully allow the directory
deletion to fail if it was reclaimed.

Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>