util/fscache: ignore entries without info-file
Extend the cache-load operation to ignore entries that lack an
object-info file. This file is written after an entire entry is
committed, and thus can be easily used to identify corrupted entries.

Note that proper corruption detection will need store-operations to
fsync() their data before writing the info-file. However, this is left
to future patches, likely together with an inline copy-tree operation,
since fsync() needs to be called recursively on each dir and file.

Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>