util/fscache: store watermarks in cache-info
Store a new high and low watermark in the cache-info. Both are specified
in percent of cache-usage and thus clamped to [0, 100].

The high-watermark specifies at which point cache reclaim should start.
That is, once cache-usage exceeds the high-watermark, cache reclaim will
kick in and drop old entries.

The low-watermark specifies at which point cache reclaim should stop.
This means, once cache reclaim kicks in, it will reclaim space until the
low watermark is reached.

Note that neither of the watermarks is used, yet. This merely adds the
information to the cache-info file and verifies the parser.

Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>