util/fscache: update mtime on cache-hits
Update the modification-time of cache entries whenever a lookup is
successful. This will be used in the future to get better LRU-based
cache-reclaim and avoid reclaiming hot cache entries.

Note that we explicitly do not use `atime`, since many mount-options
prohibit explicit updates of `atime`. Furthermore, we do not want any
cache-maintenance operations to update the `atime` accidentally, so
`mtime` seems like a good fit.

Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>