Revert "stages/mkdir: explicitly set mode using `os.chmod`"
This reverts commit a988aacf996b41c0367f571d7e526045e0353102.

After some discussion, the original behavior was intentional. With the
added support for gracefully handling the existence of directories, the
stage would originally not set the mode of an existing directory, while
now it will. Additional issue is that `mkdir` applies the provided mode
- umask, which was intentional. Setting the same mode without taking
umask value into account is not desired.